+++ dev/frames/spectrometer_frame.py
@@ -1,8 +1,7 @@
 from dev.images.images import *
+from dev.frames.notification import *
 import customtkinter as ctk
-import numpy as np
 import matplotlib.pyplot as plt
-from seabreeze.spectrometers import Spectrometer
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2Tk
 import tkinter
 import csv
@@ -11,18 +10,19 @@ import csv
 class SpectrometerFrame(ctk.CTkFrame):
     def __init__(self, master, spectrometer):
         super().__init__(master)
-        self.grid_columnconfigure(0, weight=1)
-        self.grid_rowconfigure(4, weight=1)
+        self.grid_columnconfigure(1, weight=1)
+        self.grid_rowconfigure(3, weight=1)
+        self.grid_propagate(False)
 
         if not spectrometer:
             self.label = ctk.CTkLabel(self, text="No spectrometer", font=("Arial", 25))
-            self.label.grid(row=4, column=0, padx=5, pady=5)
+            self.label.grid(row=3, column=1, padx=5, pady=5)
             return
 
         self.disconnected_label = ctk.CTkLabel(self, text="Disconnected", font=("Arial", 25))
-        self.disconnected_label.grid(row=4, column=0, padx=5, pady=5)
+        self.disconnected_label.grid(row=3, column=1, padx=5, pady=5)
         self.disconnected_button = ctk.CTkButton(self, text="", width=30, height=40,  image=img_retry, fg_color="transparent", command=self.try_connection)
-        self.disconnected_button.grid(row=4, column=0, padx=5, pady=(75, 0))
+        self.disconnected_button.grid(row=3, column=1, padx=5, pady=(75, 0))
 
         self.wavelengths = None
         self.intensities = None
@@ -40,9 +40,9 @@ class SpectrometerFrame(ctk.CTkFrame):
 
             self.figure, self.plot1 = plt.subplots(figsize=(6, 4), dpi=100)
             self.canvas = FigureCanvasTkAgg(self.figure, master=self)
-            self.canvas.get_tk_widget().grid(row=1, column=0, sticky="nsew")
+            self.canvas.get_tk_widget().grid(row=0, column=1, sticky="nsew", rowspan=4)
             toolbar_frame = tkinter.Frame(self)
-            toolbar_frame.grid(row=1, sticky="new")
+            toolbar_frame.grid(row=0, column=1, sticky="new")
             self.toolbar = NavigationToolbar2Tk(self.canvas, toolbar_frame)
             self.toolbar.update()
 
@@ -51,9 +51,9 @@ class SpectrometerFrame(ctk.CTkFrame):
             self.play_button = ctk.CTkButton(self, text="", width=30, height=40,  image=img_play, fg_color="transparent", command=self.start_spectrometer)
             self.play_button.grid(row=0, column=0, padx=5, pady=5, sticky="nw")
             self.pause_button = ctk.CTkButton(self, text="", width=30, height=40, image=img_pause, fg_color="transparent", command=self.stop_spectrometer)
-            self.pause_button.grid(row=0, column=0, padx=(50,0), pady=5, sticky="nw")
+            self.pause_button.grid(row=1, column=0, padx=5, pady=5, sticky="nw")
             self.save_button = ctk.CTkButton(self, text="", width=30, height=40, image=img_save, fg_color="transparent", command=self.save_data)
-            self.save_button.grid(row=0, column=0, padx=(95,0), pady=5, sticky="nw")
+            self.save_button.grid(row=2, column=0, padx=5, pady=5, sticky="nw")
 
             if self.spectrometer.enable:
                 self.start_spectrometer()
@@ -89,6 +89,8 @@ class SpectrometerFrame(ctk.CTkFrame):
                         writer = csv.writer(file)
                         writer.writerow(['Wavelength [nm]', 'Intensity [counts]'])
                         writer.writerows(zip(self.wavelengths, self.intensities))
+                    self.master.notification(f"Successfully saved as", file_path, "#1a8300")
                     print(f"Data saved successfully to {file_path}")
                 except Exception as e:
                     print(f"Error saving data: {e}")
+                    self.master.notification(f"Impossible to save as", file_path, "#8e0101")