+++ main.py
@@ -17,5 +17,5 @@ lm_microscope = MyMicroscope("Lisa Meitner",
     lm_filter_12 = MyFilter("Filter 12", "xxxx"),
     lm_platform = MyPlatform("Platform", "xxxx"))
 
-app = MyApp("V0.1.4", (1200, 700), gm_microscope, lm_microscope)
+app = MyApp("V0.2.0", (1200, 700), gm_microscope, lm_microscope)
 app.mainloop()